Docker-microservice/web/requirements.txt
flask
gunicorn
psycopg2-binary
redis